Issue #138: Moved Mock* classes to src/test/java.
